gst/rtpmanager/rtpsource.*: When no payload was specified on the caps but there was a clock-rate, assume the clock-ra...
Original commit message from CVS:
* gst/rtpmanager/rtpsource.c: (rtp_source_init),
(rtp_source_update_caps), (get_clock_rate):
* gst/rtpmanager/rtpsource.h:
When no payload was specified on the caps but there was a clock-rate,
assume the clock-rate corresponds to the first payload type found in the
RTP packets. Fixes #565509.

+++ gst/rtpmanager/rtpsource.h
@@ -128,7 +128,7 @@ struct _RTPSource {
   gboolean      have_rtcp_from;
   GstNetAddress rtcp_from;
 
-  guint8        payload;
+  gint          payload;
   GstCaps      *caps;
   gint          clock_rate;
   gint32        seqnum_base;
